+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # 2.1.1 (2023-01-16)
 - feat: support with_cycles for get_block and get_block_by_number rpc
+- feat: support packed rpcs
+- feat: support indexer exact search mod
 
 ## 🚀 Features
 # 2.1.0 (2022-12-26)